makes bash 3 compatible
this tries to minimize the dependencies on 'modern' bash and removes egrep

instead of searching for all files and using egrep to pass the ones we want, this introduces snyk_gen_file_list which creates a search string explicitly matching the manifest files for each project type we want to scan. we also add an explicit block of any .* files or folders, as we want to traverse or search something like a .git directory either.

This also gets away from using associative arrays (which won't work in bash3) instead doing some variable expansion around vars that all have the prefix of projectTypes_* not as nice as named array elements, but should be more compatible in someones build environments